Allow to define services that use external images
