+++ dmake/core.py
@@ -5,7 +5,7 @@ import uuid
 
 import dmake.common as common
 from dmake.common import DMakeException, SharedVolumeNotFoundException
-from dmake.deepobuild import DMakeFile, append_command
+from dmake.deepobuild import DMakeFile
 
 tag_push_error_msg = "Unauthorized to push the current state of deployment to git server. If the repository belongs to you, please check that the credentials declared in the DMAKE_JENKINS_SSH_AGENT_CREDENTIALS and DMAKE_JENKINS_HTTP_CREDENTIALS allow you to write to the repository."
 
@@ -795,19 +795,19 @@ def make(options, parse_files_only=False):
     # Generate the list of command to run
     common.logger.info("Generating commands...")
     all_commands = []
-    append_command(all_commands, 'env', var = "REPO", value = common.repo)
-    append_command(all_commands, 'env', var = "COMMIT", value = common.commit_id)
-    append_command(all_commands, 'env', var = "BUILD", value = common.build_id)
-    append_command(all_commands, 'env', var = "BRANCH", value = common.branch)
-    append_command(all_commands, 'env', var = "DMAKE_TMP_DIR", value = common.tmp_dir)
+    common.append_command(all_commands, 'env', var = "REPO", value = common.repo)
+    common.append_command(all_commands, 'env', var = "COMMIT", value = common.commit_id)
+    common.append_command(all_commands, 'env', var = "BUILD", value = common.build_id)
+    common.append_command(all_commands, 'env', var = "BRANCH", value = common.branch)
+    common.append_command(all_commands, 'env', var = "DMAKE_TMP_DIR", value = common.tmp_dir)
     # check DMAKE_TMP_DIR still exists: detects unsupported jenkins reruns: clear error
-    append_command(all_commands, 'sh', shell = 'dmake_check_tmp_dir')
+    common.append_command(all_commands, 'sh', shell = 'dmake_check_tmp_dir')
 
     for stage, commands in ordered_build_files:
         if len(commands) == 0:
             continue
 
-        append_command(all_commands, 'stage', name = stage, concurrency = 1 if stage == "Deploying" else None)
+        common.append_command(all_commands, 'stage', name = stage, concurrency = 1 if stage == "Deploying" else None)
 
         stage_commands = []
         for node, order in commands:
@@ -842,22 +842,21 @@ def make(options, parse_files_only=False):
                 sys.exit(1)
 
             if len(step_commands) > 0:
-                append_command(stage_commands, 'echo', message = '- Running %s' % (display_command_node(node)))
+                common.append_command(stage_commands, 'echo', message = '- Running %s' % (display_command_node(node)))
                 stage_commands += step_commands
 
         # GPU resource lock
         # `common.need_gpu` is set during Testing commands generations: need to delay adding commands to all_commands to create the gpu lock if needed around the Testing stage
         lock_gpu = (stage == "Running App") and common.need_gpu
         if lock_gpu:
-            append_command(all_commands, 'lock', label='GPUS')
+            common.append_command(all_commands, 'lock', label='GPUS')
 
         all_commands += stage_commands
 
         if lock_gpu:
-            append_command(all_commands, 'lock_end')
+            common.append_command(all_commands, 'lock_end')
 
-        append_command(all_commands, 'stage_end')
-
+        common.append_command(all_commands, 'stage_end')
 
     # Check stages do not appear twice (otherwise it may block Jenkins)
     stage_names = set()
@@ -871,7 +870,7 @@ def make(options, parse_files_only=False):
 
     # If not on Pull Request, tag the commit as deployed
     if common.command == "deploy" and not common.is_pr:
-        append_command(all_commands, 'git_tag', tag = get_tag_name())
+        common.append_command(all_commands, 'git_tag', tag = get_tag_name())
 
     # Generate output
     if common.is_local:
