+++ dmake/python_3x.py
@@ -1,23 +1,24 @@
 from io import StringIO
+from abc import ABC, ABCMeta
 
 from collections import OrderedDict
 
 from functools import lru_cache
 
-class MetaYAML2PipelineSerialize(type):
+class MetaSerializerMixin(ABCMeta):
     @classmethod
     def __prepare__(metacls, name, bases, **kwargs):
         return OrderedDict()
     def __new__(cls, name, bases, namespace, **kwargs):
-        result = type.__new__(cls, name, bases, dict(namespace))
+        result = ABCMeta.__new__(cls, name, bases, dict(namespace))
         ns = []
         for b in bases:
-            if isinstance(b, MetaYAML2PipelineSerialize):
+            if isinstance(b, MetaSerializerMixin):
                 ns += b.__fields_order__
         result.__fields_order__ = tuple(ns) + tuple(namespace)
         return result
 
-class BaseYAML2PipelineSerializer(object, metaclass = MetaYAML2PipelineSerialize):
+class SerializerMixin(ABC, metaclass=MetaSerializerMixin):
     pass
 
 def is_string(x):
